requirements.txt
asgiref==3.7.2
Django==5.0
django-macaddress==1.8.0
netaddr==0.9.0
python-dotenv==1.0.0
sqlparse==0.4.4
tzdata==2023.3
